# server/swagger.yaml
openapi: 3.0.0
info:
  title: Agentic Schedule API
  version: 1.0.0
  description: API documentation for Contracts, Staff, Skills, DayOffRequests, and ShiftOffRequests
paths:
...
  /staff:
    post:
      summary: Create one or multiple staff members
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Staff'
                - type: array
                  items:
                    $ref: '#/components/schemas/Staff'
            examples:
              SingleStaff:
                summary: Single staff object
                value:
                  name: "John Doe"
                  contractId: 1
                  skillId: 2
              MultipleStaff:
                summary: List of staff objects
                value:
                  - name: "Jane Doe"
                    contractId: 2
                    skillId: 3
                  - name: "Alice"
                    contractId: 1
                    skillId: 2
      responses:
        '201':
          description: Staff created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 201
                  success_message:
                    type: string
                    example: Staff created successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      staffIds:
                        type: array
                        items:
                          type: integer
                        example: [1, 2, 3]
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
    get:
      summary: Get all staff
      responses:
        '200':
          description: List of staff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardStaffListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
    delete:
      summary: Delete staff by list of StaffID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                staffIds:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3]
      responses:
        '200':
          description: Staff deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 200
                  success_message:
                    type: string
                    example: Staff deleted successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      deletedCount:
                        type: integer
                        example: 2
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
  /staff/{id}:
    get:
      summary: Get staff by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Staff member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardStaffResponse'
        '404':
          description: Staff not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
  /skills:
    post:
      summary: Create one or multiple skills
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Skill'
                - type: array
                  items:
                    $ref: '#/components/schemas/Skill'
            examples:
              SingleSkill:
                summary: Single skill object
                value:
                  skillName: "Nurse"
              MultipleSkills:
                summary: List of skill objects
                value:
                  - skillName: "Doctor"
                  - skillName: "Technician"
      responses:
        '201':
          description: Skills created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 201
                  success_message:
                    type: string
                    example: Skills created successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      skillIds:
                        type: array
                        items:
                          type: integer
                        example: [1, 2, 3]
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '409':
          description: Skill already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
    get:
      summary: Get all skills
      responses:
        '200':
          description: List of skills
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSkillsListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
    delete:
      summary: Delete skills by list of SkillID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                skillIds:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3]
      responses:
        '200':
          description: Skills deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 200
                  success_message:
                    type: string
                    example: Skills deleted successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      deletedCount:
                        type: integer
                        example: 2
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
...
components:
  schemas:
...
    Staff:
      type: object
      properties:
        StaffID:
          type: integer
        Name:
          type: string
        ContractID:
          type: integer
        SkillID:
          type: integer
    Skill:
      type: object
      properties:
        SkillID:
          type: integer
        SkillName:
          type: string
...
    StandardErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        statusCode:
          type: integer
          example: 400
        success_message:
          type: string
          example: ""
        error_message:
          type: string
          example: "Invalid data"
        payload:
          type: object
          example: {}
...
    StandardStaffListResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        statusCode:
          type: integer
          example: 200
        success_message:
          type: string
          example: Staff fetched successfully
        error_message:
          type: string
          example: ""
        payload:
          type: object
          properties:
            staff:
              type: array
              items:
                $ref: '#/components/schemas/Staff'
    StandardStaffResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        statusCode:
          type: integer
          example: 200
        success_message:
          type: string
          example: Staff fetched successfully
        error_message:
          type: string
          example: ""
        payload:
          type: object
          properties:
            staff:
              $ref: '#/components/schemas/Staff'
    StandardSkillsListResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        statusCode:
          type: integer
          example: 200
        success_message:
          type: string
          example: Skills fetched successfully
        error_message:
          type: string
          example: ""
        payload:
          type: object
          properties:
            skills:
              type: array
              items:
                $ref: '#/components/schemas/Skill'
...